Configure entity relations + adjust controllers and repositories

+++ Teza/Controllers/CollectionController.cs
@@ -25,16 +25,84 @@ namespace Teza.Controllers
         }
 
         [HttpGet]
-        public async Task<ActionResult<object>> GetAllCollections()
+        public async Task<ActionResult<object>> GetAllCollectionsAsync()
         {
             try
             {
-                var allCollections = await _unitOfWork.CollectionRepository.GetAll().ToListAsync();
+                var allCollections = await _unitOfWork.CollectionRepository.GetAllCollectionsAsync();
 
                 return new SuccessModel
                 {
                     Data = allCollections,
                     Message = "Collections retrieved",
+                    Success = true
+                };
+            }
+            catch (Exception e)
+            {
+                return new ErrorModel
+                {
+                    Error = e.Message,
+                    Success = false
+                };
+            }
+        }
+
+        [HttpGet]
+        public async Task<ActionResult<object>> GetAllCollectionsByWorkspaceIdAsync(Guid workspaceId)
+        {
+            try
+            {
+                //var workspace = await _unitOfWork.WorkspaceRepository.GetWorkspaceByIdAsync(workspaceId);
+
+                //if (workspace is null)
+                //{
+                //    return new ErrorModel
+                //    {
+                //        Error = "There's no workspace with such an ID",
+                //        Success = false
+                //    };
+                //}
+
+                var allCollections = await _unitOfWork.CollectionRepository.GetByCondition(x => x.WorkspaceId.Equals(workspaceId)).ToListAsync();
+
+                return new SuccessModel
+                {
+                    Data = allCollections,
+                    Message = "Collections retrieved",
+                    Success = true
+                };
+            }
+            catch (Exception e)
+            {
+                return new ErrorModel
+                {
+                    Error = e.Message,
+                    Success = false
+                };
+            }
+        }
+
+        [HttpGet]
+        public async Task<ActionResult<object>> GetCollectionByIdAsync(Guid collectionId)
+        {
+            try
+            {
+                var collection = await _unitOfWork.CollectionRepository.GetCollectionByIdAsync(collectionId);
+
+                if (collection is null)
+                {
+                    return new ErrorModel
+                    {
+                        Error = "There's no collection with such an ID",
+                        Success = false
+                    };
+                }
+
+                return new SuccessModel
+                {
+                    Data = collection,
+                    Message = "Collection retrieved",
                     Success = true
                 };
             }
@@ -54,6 +122,21 @@ namespace Teza.Controllers
             try
             {
                 _unitOfWork.CollectionRepository.Create(collection);
+
+                var workspace = await _unitOfWork.WorkspaceRepository.GetWorkspaceByIdAsync(collection.WorkspaceId);
+                
+                if (workspace is null)
+                {
+                    return new ErrorModel
+                    {
+                        Error = "There's no workspace with such an ID",
+                        Success = false
+                    };
+                }
+
+                workspace.Collections.Add(collection);
+                //collection.Workspace.Collections.Add(collection);
+
                 await _unitOfWork.SaveChangesAsync();
 
                 return new SuccessModel
@@ -100,12 +183,36 @@ namespace Teza.Controllers
         }
 
         [HttpDelete]
-        public async Task<ActionResult<object>> Delete(int id)
+        public async Task<ActionResult<object>> Delete(Guid collectionId)
         {
             try
             {
-                Collection collection = _unitOfWork.CollectionRepository.GetById(id);
+                var collection = await _unitOfWork.CollectionRepository.GetCollectionByIdAsync(collectionId);
+
+                if (collection is null)
+                {
+                    return new ErrorModel
+                    {
+                        Error = "There's no collection with such an ID",
+                        Success = false
+                    };
+                }
+
                 _unitOfWork.CollectionRepository.Delete(collection);
+
+                var workspace = await _unitOfWork.WorkspaceRepository.GetWorkspaceByIdAsync(collection.WorkspaceId);
+
+                if (workspace is null)
+                {
+                    return new ErrorModel
+                    {
+                        Error = "There's no workspace with such an ID",
+                        Success = false
+                    };
+                }
+
+                //workspace.Collections.Remove(collection);
+
                 await _unitOfWork.SaveChangesAsync();
 
                 return new SuccessModel
